fixing bug in sort.f due to typecasting large integers

+++ src/common/sort.f
@@ -82,7 +82,7 @@ C===================================================================
  
       INTEGER   LSTK(31),RSTK(31),ISTK
       INTEGER   L,R,I,J,P,INDEXP,INDEXT
-      REAL      DATAP
+      integer      DATAP
  
 C     QuickSort Cutoff
 C
